xhs_crawler/requirements.txt
mysql-connector-python
pandas
openpyxl
DrissionPage
logging
tqdm